+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.5.9 (not released yet)
+## 1.6.0 (not released yet)
 
 - updated `PyICU` dependency so it installs smoothly on distros with new sw versions (eg. Gentoo & Arch)
 - updated `cryptography` and `pyOpenSSL` to make things work with new OpenSSL
@@ -10,6 +10,7 @@
 
 - New config option `web.paths` for fetch other URLs than just website root
 - Workaround for https://github.com/pyca/cryptography/issues/3856
+- Fix for webservers redirecting to invalid labels (eg. `HTTP 301, Location: https://www..cz/` → `urllib3.exceptions.LocationParseError`)
 
 ## 1.5.8 (2021-03-23)